correct type

+++ BackendService/Models/FoundProductDetails.cs
@@ -3,7 +3,7 @@ namespace BackendService.Models;
 public class FoundProductDetails
 {
     public int ProductId { get; set; }
-    public int Lat { get; set; }
-    public int Long { get; set; }
+    public double Lat { get; set; }
+    public double Long { get; set; }
 
 }
